devman_lessons/telegram_bot/requirements.txt
python-dotenv==0.20.*
python-telegram-bot==13.*
pytimeparse==1.1.*